+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * #29 dot-slash paths do not work with default Jetty servlet
 * #30 Servlets/Filters: Support for 'urlPatterns' mapping in YAML
 * #31 Support for servlet and filter parameters configured in annotations 
+* #34 Configurable request and response header size
 
 ## 0.14: